Migrate to go1.17, remove inserted and deleted bitmaps (#20)
This PR removes inserted and deleted bitmaps previously stored in the transaction. This simplifies a lot of things, as now the Commit struct is as a simple array of buffers and a dirty bitmap. It also makes Delete() function on the column implementation redundant, since we can now simply switch on the operation type during Apply().

I also migrated and tested the library on go 1.17, which resulted in increased performance.