Add partial snapshotting support (#36)
This PR adds WriteTo and ReadFrom methods on the collection as a second step to add durability. This allows to encode/decode an entire collection. In its current form, all write transactions need to be stopped prior to calling WriteTo() method, otherwise the snapshot might be in inconsistent state (partially written transactions).

Given that, this change is still does not provide full durability for the data store, given that I still need to implement redo logging (wal) and make sure that the restore can re-build a consistent state of the collection.