Chunked snapshot writer/reader (#37)
This PR is a second work package required in order to properly implement snapshotting properly. This changes the way snapshots are being written and read, which can now be done chunk by chunk, allowing for simultaneous write transactions to go through.

Once this is merged, next step would be to implement a redo log and write into the commit all of the changed chunks that are happening during the snapshot process itself, allowing us to restore a consistent state of the database on loading.