Add snapshot and restore methods (#39)
This PR adds `Snapshot()` and `Restore()` methods to a collection. In order to be able to perform snapshots while writes are happening, we need to also store the commits. This PR adds an encoder/decoder for an entire commit which consists of a chunk and a series of update buffers.